apps/langchain-api/API/requirements.txt
# pip install -r requirements.txt 

langchain
langchain-openai
python-dotenv
langchain_core
beautifulsoup4
faiss-cpu


langchain_community
openpyxl
python-docx
PyPDF2
docx2txt
unstructured
networkx
pandas

flask